Address comments and split files
